BuildActions

Documentation

type SelectElem = Elem'

This process for processing forms to URLs is done by creating 4 different flows

1 =: 1 ; single valid value for the eventual key ;; typically for verification purposes | 2. 1 =: { Set of Strings } ; we have some amount of Text inputs which allow arbitrary input ;; --> we use this to skew search results to our best efforts through info on domain/keywords | 3 -- A) Select-like Elements | -- B) Radio Elements

3A & 3B result in same type eventually -> Map (Namespace/Key) [Option] optimal solution begins generating "stream" of urls on new instance of (2)

In our business logic, it revolves around 2. ; we have maybe no way of knowing for sure which | textinput is which and how they interrelate | -> (eg. CASE1: 1st is paper title, 2nd is author | CASE2: Only 1 text input for any)

We assume 1st is most likely but if it fails or gives unusual responses then we go to next | But for lack of free development time we are hoping this just works the first time and doesn't | need a ton of attention/error handling

type Namespace = Text

Name and Namespace are really same shit; might just converge | Refer to literally "name" attribute

type Option = Text

This is an operationally focused type where | a certain namespace is found to have n num of Options

data QParams

Constructors

Opt (Map Namespace [Option])

SimpleKV (Text, Text)

type ReqBody = String

type FormOptionRadio = Map String [String]

I believe this could instead compose Name/Namespace | OR!! Is the [String] the list of Options?

type FormOptionRadio' = Map Namespace [Option]

getContainedUrls :: Elem' a -> Maybe [String]

inputTypesCondensed :: [a]

data FormRes

Constructors

FormLinks [Text]

type HrefsOfPage = [Text]

since it can have multiple forms | This is more for my understanding of the bigger picture of links management if say: -> we found ALL forms and created ALL links but this would be full of trash ofc

A key goal of the link management system is to allow for precedence ; ie we've found browsing | page but prefer search page therefore: we continue to look for hrefs in such a case until we've exhausted the basic site tree

data HrefsSite

Constructors

HrefsSite [HrefsOfPage]

type URL = Text

data PageLinker'

Constructors

PL URL [FormRes] [HrefsSite]

finish :: [String] -> [String] -> String

finish' :: ([String], [String]) -> Int -> String -> String

finish'' :: [Text] -> [Text] -> QueryString

finish''' :: ([Text], [Text]) -> Int -> Text -> QueryString

genTInputOpts' :: Text -> ([Text], [Text]) -> [QueryString]

Basically just calls finish' with starting state

genSerchStrm :: String -> ([String], [String]) -> [Text]

Basically just calls finish' with starting state

allElems :: [String]

inputElems :: Maybe [String]

input has attr: type="" which determines value type for form url | Attrs that I care about really are refs@( name || id ) , (Maybe) value, type,

inputTypes :: [String]

data InputElem

Seems sepElems is not yet perfected for the whole scope of HTML Possibilities but | All functions that pull from (,,,) / 4[]

Could we just use parseElemHead here? probably gonna be waaaaay more efficient with self-closing tags and I dont think there's any relevance of the (maybe innerText) | && shouldnt this specify certain attrs? 145

Constructors

Radio Namespace Option

SelectElem (Namespace, [Option])

Basic Namespace (Maybe Option)

data ParsedForm

Constructors

ParsedForm Action Method [InputElem]

type SepdStruct = ([(Namespace, Option)], [(Namespace, [Option])], [(Namespace, Option)], [Namespace])

(Radio | Var | Basic | TInput)

sepElems' :: [InputElem] -> SepdStruct -> SepdStruct

invalidSearchElems :: [Maybe InputElem] -> ParsecT s u m [InputElem]

searchForm :: Stream s m Char => ParsecT s u m (Elem' String)

Scape pattern that matches on search forms

formElem :: Stream s m Char => ParsecT s u m ParsedForm

Scape pattern that matches on search forms as well as extracting the action and method attributes

inputElem :: Stream s m Char => ParsecT s u m (Maybe InputElem)

Scrape pattern for input element

radioBasic :: Stream s m Char => ParsecT s u m InputElem

radioBasic' :: Stream s m Char => ParsecT s u m (Maybe InputElem)

parserZero means we dont care, Nothing means this form is invalid

processInputEl :: Attrs -> ParsecT s u m (Maybe InputElem)

Handles differences in implications from all input types | When an input given its type, would not impact the query url, it's discarded

handleNumber :: Attrs -> ParsecT s u m (Maybe InputElem)

selectEl :: Stream s m Char => ParsecT s u m InputElem

formRaw <- elemParser (Just ["form"]) (Just $ inputElem) [] | return ParsedForm Action $ build (matches formRaw)...

Matches on

optionParser :: Stream s m Char => ParsecT s u m String

innerFormParser :: Stream s m Char => ParsecT s u m [Elem' String]

type TInputOpt = QueryString

type Action = Text

type BaseUrl = Url

data FilledForm

Constructors

FilledForm

Fields

actionUrl :: Url

reqMethod :: Method

from action attribute

searchTerm :: Term

textInputOpts :: [TInputOpt]

qStringVariants :: [QueryString]

data SearchSumm

If I instead just pass a baseUrl to buildFormSummary then I can return as:

Constructors

SearchSumm Term Method Action [TInputOpt] [QueryString]

type Term = String

IMPLEMENTATION | *** MAIN of forms / search mkSearchEnum :: Url -> Elem' a -> Term -> Either FormError SearchEnum mkSearchEnum baseUrl form term = do (FilledForm baseUrl reqMethod term aAttr searchQueriesTextOpts searchQueriesEnumd) <- buildFormSummary baseUrl term form mkSearch' baseUrl reqMethod term aAttr (head searchQueriesTextOpts) searchQueriesEnumd

data SearchEnum

Constructors

SearchEnum Term [Request]

mkSearch' :: Url -> Method -> Term -> Action -> QueryString -> [QueryString] -> Either FormError SearchEnum

showQString :: [(Namespace, Option)] -> Text

type SearchQuery = Query

type Query = Text

data FormError

Constructors

InvalidElement

ParsecError ParseError

UrlError

mkFormInputs :: String -> SepdStruct -> ([TInputOpt], [QueryString])

(Radio | Var | Basic | TInput)

findForms :: String -> Maybe [ParsedForm]

NOTE: formElem is a specialized

fillForm :: ParsedForm -> Url -> String -> FilledForm

mkFilledForm :: Url -> String -> Elem' String -> Either FormError FilledForm

Lazily build FilledForm so that the full list of possible querystrings is not evaluated

mkQParams :: String -> [InputElem] -> ([TInputOpt], [QueryString])

actionAttr :: Elem' a -> Text

mkBasicParams' :: [Elem' a] -> QueryString

mkSubsetVars :: [Elem' a] -> [Elem' a] -> Map Namespace [Option]

searchTermSubPaths :: Map Namespace [Option] -> QueryString -> [QueryString]

type QueryString = [(Namespace, Option)]

getAttrVal :: String -> Elem' a -> String

buildSearchUrlSubsets :: Map Namespace [Option] -> [QueryString]

With a Map of an html Namespace and all its possible defined options, lazily create all possible search queries | for a given Search term

sepElems2 :: ElementRep a => [a b] -> ([a b], [a b], [a b], [a b])

Note: this url may fail

SHOULD I? create a list of baseUrls, where each could fail and if so, next buildBaseFormUrl baseUrl generalSearchTermToSearchForTextbox ((Elem'{..}):elems) elem == "textarea" = genSearchTermTBox plus when input type attribute is text elem == "button" = ""

sepElems :: [Elem' a] -> ([Elem' a], [Elem' a], [Elem' a], [Elem' a])

data SelectElem'

above should check if : el elem == "select"

Constructors

SelectElem' Namespace [Option]

mkOptMaps :: [SelectElem a] -> [(Namespace, [Option])]

mkOptMaps' :: [(Namespace, [Option])] -> Map Namespace [Option]

for variable elems, this will be a case where we need to parse inner text again for the option elements

mkOptMaps'' :: [SelectElem a] -> Map Namespace [Option]

mkOptMapSingle :: SelectElem a -> Maybe (Namespace, [Option])

This likely has a far better implementation although im honestly not sure how much "overhead" | would be removed by not using parser again | Likely tho; -> best option always is 1 pass parsing every time we call "parse" I believe its equivalent to parse str where str = strFullHtml ++ strInnerHtmlSpec where all input is consumed and each index has N number of (->Boolean) functions as options thru |

applyFailStream :: (a -> Bool) -> [a] -> [a]

optionElemsPat :: Maybe [String]

formOptionsParser :: Stream s m Char => ParsecT s u m [String]

toLis2' :: [(Namespace, Option)] -> Namespace -> Option -> [(Namespace, Option)]

I have to use (<>) to deal with the fact that the end of the list would look the exact same as an "inner end" | but this doesnt mean that I will have infinite listing ending up with something like [[[[[[[[a]]]]]]]]

Although, ^ this may be more likely a problem if using cons, (:) since it creates a more complex type | which works fine at the last iteration where strings are thus being operated like so

Meaning that at whateverrr depth we will be able to (<>) the [string]

Meaning this will work for any depth of tree

Should always be singleton !!!

toStr2 :: Text -> Namespace -> Option -> Text

singlefOp' :: [(Namespace, Option)] -> [(Namespace, [Option])] -> [[(Namespace, Option)]]

singlefOp can be thought of as a state carrying function which evaluates to N number of | functions which evaluate to a list of the expression of toStr2

singlefOp :: Text -> [(Namespace, [Option])] -> [Text]

Is essentially the core of buildFormUrlEndings; creates all url search queries

createBasicQKVPair :: Elem' a -> (Namespace, Option)

Creates specifically last branch ^^ levels to go (below) no levels to go (up)

So maybe f takes a state parameter that is determined by the higher-level-set value

SO add the results of f :: ItemOfCurrentList -> [ItemOCL] -> [String]

for basic

createBasicQKVPairs :: [Elem' a] -> QueryString

createQKVPairWithTerm :: Text -> [Elem' a] -> [Text]

NOTE: this below function sucks createQKVP...Term | Instead i will do a cycle pattern that controls it

case request of | Fail -> move first query param to end of list/string and move the search term | Success -> well then do ya thang shawty

NOT REFERENCED

type NamespacePair = (Namespace, Option)

mkNameVal :: Elem' a -> NamespacePair

radiosToMap :: [NamespacePair] -> Map Namespace [Option]

iterRadios :: [Elem' a] -> Map Namespace [Option] -> Map Namespace [Option]

lookupOrInsertName :: Attrs -> Map Namespace [Option] -> Map Namespace [Option]

searchIsBigorSmall :: [Elem' a] -> Bool

NOT REFERENCED groupRadio :: [Elem'] -> Map Name [Option] groupRadio (elem:elems) = go elems [] where go [] listOut = listOut go (elem: elems) listOut = go' elem (opt:listOut)

Just an arbitrarily set number of elements | Could also see if select-like elems count == 0

findPagination :: ParsecT s u m (TreeHTML a)

Just use substring | Will it be this page tho? Or a sort of weird response

Should use findSomeSameEl probably

structuredBrowsingLinksExist :: ParsecT s u m (TreeHTML a)

validHrefs :: Stream s m Char => String -> ParsecT s u m (Maybe [String])

getValidHref :: Stream s m Char => String -> ParsecT s u m String

parseDrvPagination :: Stream s m Char => String -> ParsecT s u m UrlPagination

type PaginationElems = (Elem' String, Elem' String, Elem' String)

derivePagination :: Url -> PaginationElems -> Maybe UrlPagination

commonUrlTrups :: PaginationElems -> [(Url, Url, Url)]

hrefElHs :: Elem' String -> [ElemHead]

validHrefElHs :: Stream s m Char => ParsecT s u m ElemHead

commonElHUrls :: [ElemHead] -> [ElemHead] -> [ElemHead] -> [(Url, Url, Url)]

findSep :: (ElemHead -> Bool) -> [ElemHead] -> (Maybe ElemHead, [ElemHead])

Note that in use, what builds the predicate will be given in the return just not through this function

firstJust :: (a -> Maybe b) -> [a] -> Maybe b

tillWeGetItRight :: (a -> Maybe b) -> [a] -> Maybe b

funcyAf :: String -> String -> String -> String -> Maybe UrlPagination

paginationElements :: Stream s m Char => ParsecT s u m (Elem' String, Elem' String, Elem' String)

Should be replaced by a function that ensures they are the same element | If there is a union of elemHeads inside the given element then use order, we could even give a fake attribute | named position = n triZip :: [a] -> [b] -> [c] -> [(a,b,c)] triZip [] _ _ = [] triZip _ [] _ = [] triZip _ _ [] = [] triZip (a:as) (b:bs) (c:cs) = (a,b,c) : triZip as bs cs

Note! because we must have the same ElemHead in all 3, we only need to iterate through the options | afforded by the first element

catEithers :: [Either a b] -> [b]

Orphan instances

ShowHTML a => ShowHTML (Maybe a)

Instance details

Methods

showH :: Maybe a -> String