Elem.ElemHeadParse

Documentation

parseAttrSafe :: Stream s m Char => String -> ParsecT s u m String

needs to use many for multiple links

Safe because it forces parse of the entire ElemHead then pulls if there | Designed for use in findSomeHtml

hrefParser :: Stream s m Char => ParsecT s u m String

Done like this so that it reliably is true link and not false positive

parseOpeningTagF :: Stream s m Char => String -> (String -> Bool) -> ParsecT s u m ElemHead

hrefParser' :: Stream s m Char => (String -> Bool) -> ParsecT s u m String

Allows parsing with high level predicate

attrValue :: Stream s m Char => ParsecT s u m [Char]

In future should add replace of apostrophe and similar issues to corresponding html representations

attrValuesExist :: [(String, String)] -> [(String, Maybe String)] -> Bool

Both attrValue functions mimic map functionality

attrValueExists :: [(String, String)] -> (String, Maybe String) -> Bool

attrName :: Stream s m Char => ParsecT s u m String

attrParser :: Stream s m Char => ParsecT s u m (String, String)

Need lower|UPPER case insensitivity

for generalization sake

attrsParser :: Stream s m Char => [(String, Maybe String)] -> ParsecT s u m (Either AttrsError (Map String String))

isAttrsMatch' :: Map String String -> [(String, Maybe String)] -> Bool

isAttrsMatch :: Map String String -> [(String, Maybe String)] -> Bool

attrsFit :: Map String String -> [(String, String -> Bool)] -> Bool

attrsMatch' :: Map String String -> Map String String -> Bool

attrsMatch :: [(String, String)] -> Map String String -> Bool

attrsParserDesc :: Stream s m Char => [(String, String)] -> ParsecT s u m (Map String String)

parseOpeningTagDesc :: Stream s m Char => Maybe [Elem] -> [(String, String)] -> ParsecT s u m (Elem, Attrs)

digitEq :: String -> String -> Bool

Allows for certain degrees of freedom such as 1 spot off eg 123 vs 1230 (or even (109|22)) | as well as any numerical digit must also be a numerical digit from 0 to 9

saveDigitEq :: String -> String -> Bool

svDigEq :: String -> String -> Bool

digitEqFree :: [Char] -> [Char] -> Bool

unfit :: [(String, String)] -> Map String String -> [(String, String)]

mkAttrsDesc :: [(String, String)] -> [(String, String -> Bool)]

parseOpeningTag :: Stream s m Char => Maybe [Elem] -> [(String, Maybe String)] -> ParsecT s u m (Elem, Attrs)

NOTES if href="#" on form -> just means scroll to top

May rename parseOpeningTag to elemHeadParser | -> Case of input tag: ...." DONE ie no innerhtml or end tag | then this would be more efficient or even maybe we should add an option via | a datatype: InnerTextOpts a = DoesntExist --efficient parser | AnyText | ParserText a

mkElemtagParser :: Stream s m Char => Maybe [Elem] -> ParsecT s u m String

For elemsOpts, will either be | Parser: (anyChar) | Parser: (buildElemsOpts elems)

buildElemsOpts :: Stream s m Char => [Elem] -> ParsecT s u m String

FUTURE USE CASES: buildElemsOpts :: [ParsecT s u m a] -> ParsecT s u m a -- using |