Elem.ITextElemParser

Documentation

type Html = String

Note: will need more complex accumulator for case where an elem has two distinct text segements broken up | by an element, (rare case)

removeStyleTags :: Html -> Html

catEithers :: [Either e a] -> [a]

divideUp :: Stream s m Char => ParsecT s u m String -> ParsecT s u m [Either String String]

onlyPlainText :: Stream s m Char => ParsecT s u m String

data AccumITextElem a

Constructors

ACT [String]

textOnlyFoldr :: HTMLMatcher AccumITextElem String -> (String, [String]) -> (String, [String])