Elem.TreeElemParser

Documentation

data Many a

Constructors

Many a

One a

treeLookupIdx :: TreeIndex -> Forest a -> a

treeElemParser :: (Stream s m Char, ShowHTML a) => Maybe [Elem] -> Maybe (ParsecT s u m a) -> [(String, Maybe String)] -> ParsecT s u m (TreeHTML a)

Like elemParser, this matches on an html element but also represents the innerHTML | as a Tree ElemHead so that we can match this structure in elements further down in the DOM | see groupHtml and treeElemParserSpecific

selfClosing :: [String]

treeElemParser' :: (Stream s m Char, ShowHTML a) => Maybe [Elem] -> Maybe (ParsecT s u m a) -> [(String, Maybe String)] -> ParsecT s u m (TreeHTML a)

Used by treeElemParser, is not an interface, use treeElemParser

innerTreeElemParser :: (ShowHTML a, Stream s m Char) => Elem -> Maybe (ParsecT s u m a) -> ParsecT s u m (String, [a], [Tree ElemHead])

The real difference between (htmlGroup _ _ _) and specificRepetitiveForest is a matter of if we accept the next | piece to be a new discovery to match on or if we are in that process of matching what we just found

type SubTree a = [Tree a]

NOTE: In future: create function that simplifies all numbers that will be compared to their number of digits 1426674 -> 1234567 1834324 -> 1234567 (==) -> True

Ideal case is that we can do (Many a) struturing even if interspersed with text which would solve issues like | many search terms highlighted, we then wouldnt need to know what the search term is | AND!! we have already seen that it could for exmaple be: aclass="hiddenText"Hockey/b/a

treeElemParserSpecific :: (Stream s m Char, ShowHTML a) => Maybe (ParsecT s u m a) -> Elem -> [(String, String)] -> SubTree ElemHead -> ParsecT s u m (TreeHTML a)

Note: unlike other Element parsers, it does not call itself but innerParserSpecific instead loops with |

validateGPR :: [Many (Tree ElemHead)] -> ParsecT s u m [HTMLMatcher TreeHTML a]

htmlGenParserRepeat' :: (Stream s m Char, ShowHTML a) => String -> Maybe (ParsecT s u m a) -> [Many (Tree ElemHead)] -> ParsecT s u m [HTMLMatcher TreeHTML a]

Uses HTMLMatcher to collect cases of html while parsing inside of a certain element

htmlGenParserRepeat :: (Stream s m Char, ShowHTML a) => String -> Maybe (ParsecT s u m a) -> [Many (Tree ElemHead)] -> ParsecT s u m [HTMLMatcher TreeHTML a]

BUT! we need to check off our list of demanded elements so that when we parse the end tag, we can see if | the elements (ordered and parsed only in order) were all found before the end tag for htmlGenParserRepeat it can just change the passed state of [Many (Tree ElemHead)]

specificChar :: Stream s m Char => ParsecT s u m Char

NEW IDEA!!

treeElemParserSpecificContinuous :: (Stream s m Char, ShowHTML a) => Maybe (ParsecT s u m a) -> [Many (Tree ElemHead)] -> ParsecT s u m ([Many (Tree ElemHead)], TreeHTML a)

treeElemParserSpecific is an interface to this (via innerParserSpecific) | This inner function uses the Many datatype to differentiate between whether we should expect | to parse a single element with the given specs or allow for multiple of the given element specs in a row | |

treeElemParserContains :: (Stream s m Char, ShowHTML a) => Maybe (ParsecT s u m a) -> [Many (Tree ElemHead)] -> ParsecT s u m ([Many (Tree ElemHead)], TreeHTML a)

This is largely a subfunc of htmlGenParserContains | Accepts any element and if element is in the order of our checklist-of-elems, we give the tail of elems back | if the tail reaches [] before we hit the end tag then we are successful

htmlGenParserContains :: (Stream s m Char, ShowHTML a) => String -> Maybe (ParsecT s u m a) -> [Many (Tree ElemHead)] -> ParsecT s u m [HTMLMatcher TreeHTML a]

specificChar' :: Stream s m Char => Elem -> ParsecT s u m Char

innerParserContains :: (Stream s m Char, ShowHTML a) => Maybe (ParsecT s u m a) -> Elem -> SubTree ElemHead -> ParsecT s u m ([a], String, [Tree ElemHead])

similarTreeH :: (Stream s m Char, ShowHTML a) => Maybe (ParsecT s u m a) -> TreeHTML a -> ParsecT s u m (TreeHTML a)

Very similar to treeElemParserSpecific except that it allows for a new nodes in the HTML DOM tree | to exist at random as long as when we resume parsing we still find all of the branches we found in the | TreeHTML a that is given as an arg to this function

htmlGroupSimilar :: (Stream s m Char, ShowHTML a) => Maybe [Elem] -> Maybe (ParsecT s u m a) -> [(String, Maybe String)] -> ParsecT s u m (GroupHtml TreeHTML a)

Returns an entire group of highly similar elements based on their specifications such | as their innerTrees, the element tag, and attributes. | | This can be used to autonomously determine the structure of and find search result items after you've submitted a form

takeTill :: (a -> Bool) -> [a] -> [a]

tryElHeads :: (Elem, Attrs) -> [Many (Tree ElemHead)] -> ParsecT s u m ([Tree ElemHead], [Many (Tree ElemHead)])

yields how many are still worth trying

tryElHeads' :: (Elem, Attrs) -> [Many (Tree ElemHead)] -> Either String ([Tree ElemHead], [Many (Tree ElemHead)])

innerParserSpecific :: (Stream s m Char, ShowHTML a) => Maybe (ParsecT s u m a) -> Elem -> SubTree ElemHead -> ParsecT s u m ([a], String, [Tree ElemHead])

groupify :: Eq a => [Tree a] -> [Many (Tree a)] -> [Many (Tree a)]

IS THIS IN THE RIGHT ORDER OR DOES IT NEED TO BE REVERSED? | Creates a simplified set of instructions for parsing a very specific Tree structure

htmlGroup :: (Stream s m Char, ShowHTML a) => Maybe [Elem] -> Maybe (ParsecT s u m a) -> [(String, Maybe String)] -> ParsecT s u m (GroupHtml TreeHTML a)

Returns a minimum of 2 --> almost like same should be function ; same :: a -> [a] to be applied to some doc/String | note: not sure if this exists but here's where we could handle iterating names of attributes | Can generalize to ElementRep e

multiTreeElemHeadParser :: (Stream s m Char, ShowHTML a) => ParsecT s u m a -> Many (Tree ElemHead) -> ParsecT s u m [HTMLMatcher TreeHTML a]

Deprecated: use specificContinuous style functions

Build [Many (Tree ElemHead)] | Write parser that tries each case OR parses openingTag and then decides what case it fits | For set (Previous, Next) if Next is True then delete Previous parser | Next becomes Previous in future equation(s)

Fail onto plain IText (of parent element that the parser is currently in)

Inner parser of treeElemParserSpecific specificRepetitiveForest :: (Stream s m Char, ShowHTML a) => [Many (Tree ElemHead)] -> ParsecT s u m a -> ParsecT s u m [HTMLMatcher TreeHTML a] specificRepetitiveForest [] _ = return [] specificRepetitiveForest (mElHead1:mElHead2:manyElHeads) match = do -- | ysA could be == [] htmlGenParserRepeat match (manyElHeads) -- NEW ysA {-maybeNext-} <- multiTreeElemHeadParser match mElHead2 ysB <- case ysA of Just a -> specificRepetitiveForest (mElHead2:manyElHeads) match Nothing -> -- WHat happens to the rest of Many ElHeads in this case? htmlGenParserRepeat match (multiTreeElemHeadParser match mElHead1) htmlGenParserRepeat match (manyElHeads) -- NEW return (ysA <> ysB) let -- funcP :: (ShowHTML a, Stream s m Char) => ParsecT s u m [TreeHTML a] funcP = multiTreeElemHeadParser match mElHead1 y <- htmlGenParserRepeat match funcP -- this literally just allows for matching on multiple elems too -- Only applies to "specific" functions, prev: any case ys <- case y of -- Discard last parsed pattern and go to next element formula on success of y ((Element _):xs') -> specificRepetitiveForest manyElHeads match _ -> specificRepetitiveForest (manyElHead:manyElHeads) match -- return all results

This is all I actually need , no need for recursion here, since thats already done in top level func

fromMany :: Many a -> a

Is able to repeat / execute any pattern that returns multiple elements of same type |(see manyTreeElemHeadParser) htmlGenParserRepeat :: (Stream s m Char, ShowHTML a) => ParsecT s u m a -> [Many (Tree ElemHead)] -> ParsecT s u m [TreeHTML a] -- Can just apply multiTreeElemHeadParser (if i should) inside -> ParsecT s u m [HTMLMatcher TreeHTML a]

HTMLGenParserRepeat is in this use case always going to be exact ie these 3 elems then the end tag | ... and maybe some text in between there

OF the cases we can do this: | parse and repeat function/recurse | or find end tag >> return [] which ends list

stylingTags :: [Elem]

Note that multiTreeElemHeadParser is still not handled, all I need to do is auto delete if only one | actual function of multiTreeElemHeadParser will not be used but broken up

stylingElem :: Stream s m Char => ParsecT s u m String

Just gives the inners

sameTreeH :: (Stream s m Char, ShowHTML a) => Maybe (ParsecT s u m a) -> TreeHTML a -> ParsecT s u m (TreeHTML a)

Interface to find same element

htmlGenParserFlex :: (Stream s m Char, ShowHTML a) => Maybe (ParsecT s u m a) -> ParsecT s u m (HTMLMatcher TreeHTML a)

htmlGenParser :: (Stream s m Char, ShowHTML a) => ParsecT s u m a -> ParsecT s u m (TreeHTML a) -> ParsecT s u m (HTMLMatcher TreeHTML a)

specificForest :: (Stream s m Char, ShowHTML a) => [Tree ElemHead] -> ParsecT s u m a -> ParsecT s u m [HTMLMatcher TreeHTML a]

Deprecated: you likely need specificRepetitiveForest

Library function for when you want an exact match, if 3 of ElemHead A then it looks for 3 Elemhead A accumMaybe' :: [HTMLMatcher] -> ParsecT s u m a

nodeToTreeElemExpr :: (Stream s m Char, ShowHTML a) => Tree ElemHead -> ParsecT s u m a -> ParsecT s u m (TreeHTML a)

innerElemParser2 :: (ShowHTML a, Stream s m Char) => String -> Maybe (ParsecT s u m a) -> ParsecT s u m [HTMLMatcher TreeHTML a]

Used by treeElemParser'

treeElemParserAnyInside :: (Stream s m Char, ShowHTML a) => Maybe (ParsecT s u m a) -> ParsecT s u m (TreeHTML a)

anyHtmlGroup :: (ShowHTML a, Stream s m Char) => ParsecT s u m (GroupHtml TreeHTML a)

findAllSpaceMutExGroups :: (ShowHTML a, Stream s m Char) => ParsecT s u m (Maybe [GroupHtml TreeHTML a])

findAllMutExGroups' :: a