Elem.Types

Documentation

class HtmlMatcher a

class ShowHTML a where

Methods

showH :: a -> String

class ElementRep (a :: * -> *) where

Methods

elTag :: a b -> Elem

attrs :: a b -> Attrs

innerText' :: a b -> String

matches' :: a b -> [b]

class (ShowHTML c, ElementRep a) => InnerHTMLRep (a :: * -> *) (b :: * -> *) c | a c -> b c where

Methods

foldHtmlMatcher :: [HTMLMatcher a c] -> b c

innerText :: b c -> String

matches :: b c -> [c]

data TreeHTML a

Note, this is the representation i'll be using

Constructors

TreeHTML

Fields

_topEl :: Elem

_topAttrs :: Map String String

_matches' :: [a]

_innerText' :: String

_innerTree' :: Forest ElemHead

data InnerTextHTMLTree a

Constructors

InnerTextHTMLTree

Fields

_matches :: [a]

_innerText :: String

innerTree :: Forest ElemHead

data Elem' a

node-like

Constructors

Elem'

Fields

_el :: Elem

_attrs :: Map String String

innerMatches :: [a]

innerHtmlFull :: String

data InnerTextResult a

Constructors

InnerTextResult

Fields

_matchesITR :: [a]

_fullInner :: String

type HMatcher' a b c = [HTMLMatcher b c] -> a c

data HTMLMatcher (a :: * -> *) b

Constructors

IText String

Element (a b)

Match b

type HTMLMatcherM a = HTMLMatcher TreeHTML a

type Inner a = HTMLMatcher Elem' a

type HTMLMatcherList a = HTMLMatcher [] a

data HTMLBare e a

Constructors

HTMLBare

Fields

tag :: Elem

attrsss :: Attrs

htmlM :: [HTMLMatcher e a]

type ElemHead = (Elem, Attrs)

type Attrs = Map String String

type Elem = String

data AttrsError

Constructors

IncorrectAttrs

data MessyTree a b

At end will be able to do Eq2 of trees where tree params are (tag,attrs) | Need some "Flexible Equality" match

Would treeElemParser fail on cases like input with no end tag?

Constructors

Noise b

Nodee a [MessyTree a b]

data MessyTreeMatch a b c

Constructors

Noise' a

Match' b

Node' c [MessyTreeMatch a b c]

type TreeIndex = [Int]

data GroupHtml element a

Constructors

GroupHtml [element a] Glength MaxLength

type Glength = Int

type MaxLength = Int

ungroup :: ElementRep e => GroupHtml e a -> [e a]

mkGH :: ElementRep e => [e a] -> GroupHtml e a

longestElem :: [Elem' a] -> Maybe (Elem' a)

maxLength :: [[a]] -> [a]

biggestHtmlGroup :: [GroupHtml e a] -> GroupHtml e a

biggestGroup :: ElementRep e => [GroupHtml e a] -> GroupHtml e a

getHref :: ElementRep e => e a -> Maybe String

elemToStr :: Elem' a -> String

treeElemToStr :: ShowHTML a => TreeHTML a -> String

foldFuncMatchlist :: (ShowHTML a, ElementRep e) => HTMLMatcher e a -> InnerTextResult a -> InnerTextResult a

foldFuncTup :: (ShowHTML (e a), ShowHTML a, ElementRep e) => HTMLMatcher e a -> (String, [a]) -> (String, [a])

Bug found: matches start right way then get reversed

foldFuncTrup :: ShowHTML a => HTMLMatcher TreeHTML a -> (String, [a], Forest ElemHead) -> (String, [a], Forest ElemHead)

fst' :: (a, b, c) -> a

In our failed test case with the command : parse f "" "a/div/a" where f :: (Stream s m Char) => ParsecT s u m (TreeHTML String); f = treeElemParser (Just ["a"]) Nothing []

we can tell that foldFuncTrup has been called twice (we believe)

we will test how an element named "div" inside of "a" element would behave

snd' :: (a, b, c) -> b

thd' :: (a, b, c) -> c

foldFuncITR :: (ShowHTML a, ElementRep e) => HTMLMatcher e a -> InnerTextResult a -> InnerTextResult a

fHM_c :: (InnerHTMLRep TreeHTML InnerTextHTMLTree a, ShowHTML a) => HTMLMatcher TreeHTML a -> InnerTextHTMLTree a -> InnerTextHTMLTree a

makeBranch :: TreeHTML a -> Tree ElemHead

endTag :: Stream s m Char => String -> ParsecT s u m String

enoughMatches :: Int -> String -> Map String String -> (String, [a]) -> ParsecT s u m (Elem' a)

enoughMatchesTree :: Int -> String -> Map String String -> (String, [a], Forest ElemHead) -> ParsecT s u m (TreeHTML a)

selfClosingTextful :: (ShowHTML a, Stream s m Char) => Maybe (ParsecT s u m a) -> ParsecT s u m [HTMLMatcher e a]

This is valid according to my knowledge but need to see if better to maybe instead: do some text >>= (x -> char 'n' >> x) with source being RequestBody If i recall correctly, self-closing tags dont allow embedded elements

data UrlPagination

Constructors

UrlPagination String String

type Tag = String