Find

Documentation

findNaive :: Stream s m Char => ParsecT s u m a -> ParsecT s u m (Maybe [a])

This module provides an interface for getting patterns seperated by whatever in a given source | that you plan to parse

findSequential(_x) is for information rich elements such as products that should have multiple fields | that the user would like to return

Converts a parsing/scraping pattern to one which either returns Nothing | or Just a list of at least 1 element. Maybe type is used so that there is a clearer | distinction between a failed search and a successful one

findSequential :: Stream s m Char => [ParsecT s u m a] -> ParsecT s u m [Either ScrapeFail a]

findSequential2 :: Stream s m Char => (ParsecT s u m a, ParsecT s u m b) -> ParsecT s u m (a, b)

findSequential3 :: Stream s m Char => (ParsecT s u m a, ParsecT s u m b, ParsecT s u m c) -> ParsecT s u m (a, b, c)

findUntilMatch :: Stream s m Char => ParsecT s u m a -> ParsecT s u m a

Like find naive except that finishes parsing on the first match it finds in the document

find :: Stream s m Char => ParsecT s u m a -> ParsecT s u m [Either ScrapeFail a]

baseParser :: Stream s m Char => ParsecT s u m a -> ParsecT s u m (Either ScrapeFail a)

givesNothing :: Stream s m Char => ParsecT s u m (Either ScrapeFail a)

endStream :: (Stream s m t, Show t) => ParsecT s u m (Either ScrapeFail a)

findSomeHTMLNaive :: Stream s Identity Char => Parsec s () a -> s -> Maybe [a]

Just since do we really care about non matches?

findSomeHTML :: Stream s Identity Char => Parsec s () a -> s -> Either ParseError (Maybe [a])

findAllBetween :: a

My findAll' function design / runParserOnHtml use Maybe instead of Either to toss failure case [] -> Nothing

so it returns :: Maybe [a] = Just [a] | Nothing which will be beautiful for modeling at high level from scrape result to scrape result

I also really need to implement non-zero, non-ending predicate inner function | like nonZeroSep https://hackage.haskell.org/package/replace-megaparsec-1.4.4.0/docs/src/Replace.Megaparsec.html#sepCap

NOTE: I can replace manyTill_ with anyTill from Replace.Megaparsec

buildSequentialElemsParser :: ParsecT s u m [a]

Use with constructed for parsing datatype

findOnChangeInput :: ParsecT s u m (Elem' a)

to be applied to inner text of listlike elem