Links

Documentation

type PageNumber = Int

type Url = String

type HrefURI = String

type DOI = String

data ReferenceSys

Constructors

RefSys [String] [String]

type GeneratedLink = String

type Namespace = Text

Name and Namespace are really same shit; might just converge | Refer to literally "name" attribute

type Option = Text

This is an operationally focused type where | a certain namespace is found to have n num of Options

data QParams

More for show / reasoning rn .. non-optimal

Constructors

Opt (Map Namespace [Option])

SimpleKV (Text, Text)

type SiteTree = [(Bool, Text)]

Inter site urls and whether they have been checked for some pattern

type Html = String

findAdvancedSearchLinks :: ParsecT s u m [String]

maybeUsefulNewUrl :: String -> [(Url, a)] -> HrefURI -> Maybe HrefURI

Core function of module, filters for any links which point to other pages on the current site | and have not been found over the course of scraping the site yet | filters out urls like https://othersite.com and "#"

urlIsNew :: [(a, Url)] -> HrefURI -> Bool

maybeNewUrl :: [(Url, a)] -> HrefURI -> Maybe HrefURI

maybeUsefulUrl :: String -> HrefURI -> Maybe HrefURI

Filters javascript refs, inner page DOM refs, urls with query strings and those that | do not contain the base url of the host site

getLastPath :: Url -> String

newUrlList :: [Maybe Text] -> [(Bool, Text)] -> [(Bool, Text)]

usefulNewUrls :: String -> [(Url, a)] -> [Url] -> [Maybe HrefURI]

Input is meant to be right from

usefulUrls :: String -> [Url] -> [Maybe HrefURI]

numberOfQueryParamsIsZero :: String -> Maybe String