Scrape

Documentation

runScraperOnHtml :: ParsecT String () Identity a -> String -> Maybe [a]

Find all occurences of a given parsing/scraping pattern | e.g. getHtml' "https://google.ca" >>= return . runScraperOnHtml (el "a" []) , would give all a tag html elements on google.ca

runScraperInBody :: ParsecT String () Identity a -> String -> Maybe [a]

skipToInBody :: Stream s m Char => ParsecT s u m ()

runScraperOnBody :: ParsecT String () Identity a -> String -> Maybe [a]

skipToBody :: Stream s m Char => ParsecT s u m ()

runScraperOnHtml1 :: ParsecT String () Identity a -> String -> Maybe a

allLinks :: String -> ParsecT String () Identity [String]

uses maybeUsefulUrl to get all links on page pointing only to same site links

type Name = String

tableItem :: Name -> Elem' String

scrapeFirst :: Stream s m Char => ParsecT s u m a -> ParsecT s u m (Maybe a)

findCount :: Stream s m Char => ParsecT s u m a -> ParsecT s u m Int